tests: unify FdiskParser for gpt and mbr
